+++ src/rx-query-helper.ts
@@ -1,8 +1,11 @@
 import { LOGICAL_OPERATORS } from './query-planner';
 import { getPrimaryFieldOfPrimaryKey } from './rx-schema-helper';
 import type {
+    DeterministicSortComparator,
     FilledMangoQuery,
     MangoQuery,
+    MangoQuerySortDirection,
+    QueryMatcher,
     RxDocumentData,
     RxJsonSchema
 } from './types';
@@ -12,8 +15,15 @@ import {
     toArray,
     isMaybeReadonlyArray,
     parseRegex,
-    flatClone
+    flatClone,
+    objectPathMonad,
+    ObjectPathMonadFunction
 } from './plugins/utils';
+import {
+    DEFAULT_COMPARATOR as mingoSortComparator
+} from 'mingo/util';
+import { newRxError } from './rx-error';
+import { getMingoQuery } from './rx-query-mingo';
 
 /**
  * Normalize the query to ensure we have all fields set
@@ -187,4 +197,76 @@ export function normalizeQueryRegex(
         }
     });
     return ret;
+}
+
+
+/**
+ * Returns the sort-comparator,
+ * which is able to sort documents in the same way
+ * a query over the db would do.
+ */
+export function getSortComparator<RxDocType>(
+    schema: RxJsonSchema<RxDocumentData<RxDocType>> | RxJsonSchema<RxDocumentData<RxDocType>>,
+    query: FilledMangoQuery<RxDocType> | FilledMangoQuery<RxDocumentData<RxDocType>>
+): DeterministicSortComparator<RxDocType> {
+    if (!query.sort) {
+        throw newRxError('SNH', { query });
+    }
+    const sortParts: {
+        key: string;
+        direction: MangoQuerySortDirection;
+        getValueFn: ObjectPathMonadFunction<RxDocType>;
+    }[] = [];
+    query.sort.forEach(sortBlock => {
+        const key = Object.keys(sortBlock)[0];
+        const direction = Object.values(sortBlock)[0];
+        sortParts.push({
+            key,
+            direction,
+            getValueFn: objectPathMonad(key)
+        });
+    });
+    const fun: DeterministicSortComparator<RxDocType> = (a: RxDocType, b: RxDocType) => {
+        for (let i = 0; i < sortParts.length; ++i) {
+            const sortPart = sortParts[i];
+            const valueA = sortPart.getValueFn(a);
+            const valueB = sortPart.getValueFn(b);
+            if (valueA !== valueB) {
+                const ret = sortPart.direction === 'asc' ? mingoSortComparator(valueA, valueB) : mingoSortComparator(valueB, valueA);
+                return ret as any;
+            }
+        }
+    };
+
+    return fun;
+}
+
+
+/**
+ * Returns a function
+ * that can be used to check if a document
+ * matches the query.
+ */
+export function getQueryMatcher<RxDocType>(
+    _schema: RxJsonSchema<RxDocType> | RxJsonSchema<RxDocumentData<RxDocType>>,
+    query: FilledMangoQuery<RxDocType> | FilledMangoQuery<RxDocumentData<RxDocType>>
+): QueryMatcher<RxDocumentData<RxDocType>> {
+    if (!query.sort) {
+        throw newRxError('SNH', { query });
+    }
+
+    const mingoQuery = getMingoQuery(query.selector as any);
+    const fun: QueryMatcher<RxDocumentData<RxDocType>> = (doc: RxDocumentData<RxDocType>) => {
+        if (doc._deleted) {
+            return false;
+        }
+        const cursor = mingoQuery.find([doc]);
+        const next = cursor.next();
+        if (next) {
+            return true;
+        } else {
+            return false;
+        }
+    };
+    return fun;
 }